+++ tests/test_internal.py
@@ -1,8 +1,3 @@
-from warnings import filterwarnings
-from warnings import resetwarnings
-
-import pytest
-
 from werkzeug import _internal as internal
 from werkzeug.test import create_environ
 from werkzeug.wrappers import Request
@@ -34,23 +29,10 @@ def test_wrapper_internals():
     resp.response = iter(["Test"])
     assert repr(resp) == "<Response streamed [200 OK]>"
 
-    # string data does not set content length
     response = Response(["Hällo Wörld"])
     headers = response.get_wsgi_headers(create_environ())
-    assert "Content-Length" not in headers
+    assert "Content-Length" in headers
 
     response = Response(["Hällo Wörld".encode()])
     headers = response.get_wsgi_headers(create_environ())
     assert "Content-Length" in headers
-
-    # check for internal warnings
-    filterwarnings("error", category=Warning)
-    response = Response()
-    environ = create_environ()
-    response.response = "What the...?"
-    pytest.raises(Warning, lambda: list(response.iter_encoded()))
-    pytest.raises(Warning, lambda: list(response.get_app_iter(environ)))
-    response.direct_passthrough = True
-    pytest.raises(Warning, lambda: list(response.iter_encoded()))
-    pytest.raises(Warning, lambda: list(response.get_app_iter(environ)))
-    resetwarnings()